flask_ir/requirements_2.txt
face-recognition==1.3.0
pillow==12.0.0
transformers==4.57.1
setuptools==80.9.0

--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.8.0
torchvision==0.23.0